move test to hashgraph module
