+++ platform-sdk/consensus-hashgraph-impl/src/test/resources/org/hiero/consensus/hashgraph/minConsensusRelevantThresholdTest/node3/preconsensusEvents/2026-07-07T22+09+34.446147506Z_seq0_minr1_maxr5_orgn0.pces
